Revert "extensions BUGFIX support for more than 256 extensions in one statement"
Changed ABI so revert it back.
This reverts commit 42ddd0f3a4f13abed38fd735712c28c17ceb2475.

+++ src/extensions/metadata.c
@@ -55,7 +55,7 @@ int annotation_position(const void * UNUSED(parent), LYEXT_PAR parent_type, LYEX
 int
 annotation_final_check(struct lys_ext_instance *ext)
 {
-    uint16_t  i, j, c;
+    uint8_t  i, j, c;
     struct lys_module *mod;
     struct lys_submodule *submod;
     struct lys_type *type;
